Organisation
The minecraft_account file is a thorn in my side; I will be actively trying to improve its quality and efficiency before moving on to new aspects of the bot.

With this commit, I am moving some of the "ugly" code into their separate files, specifically within a new file called "responses", to make editing easier and possibly allow for better reusability in the future.